requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
requests>=2.27.0
newspaper3k==0.2.8
deep-translator==1.11.4
nltk==3.8.1
groq==0.4.2
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99
protobuf==4.25.2 